refactor(providers): split providers.ts into modular directory
The monolithic config/providers.ts (1,516 lines) has been split into:
- providers/logos.ts - logo imports and PROVIDER_LOGO_MAP
- providers/config.ts - SYSTEM_PROVIDERS_CONFIG and CHERRYAI_PROVIDER
- providers/urls.ts - PROVIDER_URLS and provider URL constants
- providers/index.ts - barrel re-exports

providers.ts now re-exports from providers/ for backward compatibility.
